# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Swagger Doc For Find-Apps Api ",
        "version": "1.0.0",
        "title": "Restaurant Apis"
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/restaurant/forgotpassword": {
            "post": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " forgotpassword",
                "description": "forgotpassword ",
                "operationId": "forgotpassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/forgotpassword"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/forgotpassword"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/restaurant/changepassword": {
            "post": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " changepassword",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "changepassword ",
                "operationId": "changepassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/changepassword"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/changepassword"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/addMenu"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/restaurant/updateRestaurant/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " updateRestaurant",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "updateRestaurant ",
                "operationId": "updateRestaurant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updaterestaurant"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/updaterestaurant"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/updaterestaurant"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/menu/getmenu": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " getmenu",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "getmenu ",
                "operationId": "getmenu",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenu"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenu"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/getmenubyid/{id}": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " getmenubyid",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "getmenubyid ",
                "operationId": "getmenubyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenubyid"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenubyid"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/updatemenu/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " updatemenu",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "updatemenu ",
                "operationId": "updatemenu",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updatemenu"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/updatemenu"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/updatemenu"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/changemenustatus/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " changemenustatus",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "changemenustatus ",
                "operationId": "changemenustatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/changemenustatus"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/changemenustatus"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/deletemenu/{id}": {
            "delete": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " deletemenu",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "deletemenu ",
                "operationId": "deletemenu",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/deletemenu"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/deletemenu"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
    },
...
    "servers": [
        {
            "url": "http://localhost:4000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "forgotpassword": {
                "type": "object",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "apache0225@gmail.com"
                    }
                }
            },
            "changepassword": {
                "type": "object",
                "required": [
                    "oldPassword",
                    "password",
                    "confirmedPassword"
                ],
                "properties": {
                    "oldPassword": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "confirmedPassword": {
                        "type": "string",
                        "example": "12345678"
                    }
                }
            },
...
            "updaterestaurant": {
                "type": "object",
                "required": [
                    "id",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "phone_number": {
                        "type": "array",
                        "example": [
                            "9996663330",
                            "7894561239"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "example": "Dominos"
                    },
                    "ar_name": {
                        "type": "number",
                        "example": "Harsh"
                    },
                    "image": {
                        "type": "file",
                        "example": "image.jpg"
                    },
                    "email": {
                        "type": "string",
                        "example": "example12@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "restaurant_owner_name": {
                        "type": "string",
                        "example": "harsh"
                    },
                    "address": {
                        "type": "string",
                        "example": "hno 34b hdwjj eeuue"
                    },
                    "street": {
                        "type": "string",
                        "example": "punjab"
                    },
                    "country": {
                        "type": "string",
                        "example": "india"
                    },
                    "restaurant_phones": {
                        "type": "array",
                        "example": [
                            {
                                "items": {
                                    "country_code": {
                                        "type": "Number"
                                    },
                                    "phone_number": {
                                        "type": "Number"
                                    }
                                }
                            }
                        ]
                    },
                    "latitude": {
                        "type": "string",
                        "example": "12.32"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "56.56"
                    },
                    "timezone": {
                        "type": "string",
                        "example": "utc 21"
                    },
                    "restaurant_bio": {
                        "type": "string",
                        "example": "dfjwbfewf hfhfewf"
                    },
                    "ar_restaurant_bio": {
                        "type": "string",
                        "example": "punewfewfefewfjab"
                    },
                    "currency": {
                        "type": "string",
                        "example": "$"
                    },
                    "vat": {
                        "type": "string",
                        "example": "15%"
                    },
                    "whats_app_notification": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "payment_enable": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "subscription_status": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "disable_order": {
                        "type": "boolean",
                        "example": "false"
                    }
                }
            },
...
            "getmenu": {
                "type": "object",
                "required": [],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "getmenubyid": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "updatemenu": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "deserts"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "adsfg"
                    },
                    "image": {
                        "type": "file",
                        "example": "deserts.jpg"
                    },
                    "desc": {
                        "type": "string",
                        "example": "dfghqwertsd"
                    },
                    "ar_desc": {
                        "type": "string",
                        "example": "sdfghwertyuert"
                    }
                }
            },
            "changemenustatus": {
                "type": "object",
                "required": [
                    "menu_status"
                ],
                "properties": {
                    "menu_status": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "deletemenu": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
...
        }
    }
}